Implement weather forecast page (still use hardcoded location)

+++ app/src/main/java/com/homework/mhafidhaziz/weatherapp/network/dto/Condition.kt
@@ -12,7 +12,7 @@ import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 data class Condition(
-    var code: Int,
+    var code: Double,
     var icon: String,
     var text: String
 ) : Parcelable
