+++ app/src/main/java/com/homework/mhafidhaziz/weatherapp/network/dto/Location.kt
@@ -15,7 +15,7 @@ data class Location(
     var country: String,
     var lat: Double,
     var localtime: String,
-    var localtime_epoch: Int,
+    var localtime_epoch: Double,
     var lon: Double,
     var name: String,
     var region: String,
